API

API

Input

EnergySystemModel.Parameters — Type.

Input indices and parameters for the model.

source

EnergySystemModel.Specs — Type.

Specifation for which constraints to include to the model. Constraints that are not specified are included by default.

Arguments

renewable_target::Bool: Whether to include renewables target constraint.

storage::Bool: Whether to include storage constraints.

ramping::Bool: Whether to include ramping constraints.

voltage_angles::Bool: Whether to include voltage angle constraints.

source

EnergySystemModel.Variables — Type.

Variable values.

source

EnergySystemModel.Objectives — Type.

Objective values.

source

EnergySystemModel.Parameters — Method.

Loads parameter values for an instance from CSV and JSON files. Reads the following files from instance_path.

indices.json with fields G, G_r, N, L, T, S

constants.json with fields kappa, C, C_bar, r

nodes/ – Time clustered data from the nodes with fields Dem_Inc, Load_mod, Max_Load, Avail_Win, Avail_Sol

1.csv

2.csv

...

technology.csv with fields cost, lifetime, M, fuel_cost_1, fuel_cost_2, r_minus, r_plus

transmission.csv with fields M, cost, dist, lifetime, C, B

storage.csv with fields xi, cost, lifetime, C, b0_1, ..., b0_n

Arguments

instance_path::AbstractString: Path to the instance directory.

source

Model

EnergySystemModel.energy_system_model — Function.

Creates the energy system model. Returns JuMP model.

Arguments

parameters::Parameters

specs::Specs

source

Output

EnergySystemModel.save_results — Function.

Save results, including parameter, variable, and objective values. Writes the following JSON files to output_path.

specs.json

parameters.json

variables.json

objectives.json

Arguments

specs::Specs

parameters::Parameters

variables::Variables

objectives::Objectives

output_path::AbstractString

source

Plotting

EnergySystemModel.plot_objective_values — Function.

Plot objective value and individual objective values.

source

EnergySystemModel.plot_generation_dispatch — Function.

Plot generation dispatch.

source

EnergySystemModel.plot_generation_capacities — Function.

Plot generation capacities.

source

EnergySystemModel.plot_transmission_flow — Function.

Plot transmission flow.

source

EnergySystemModel.plot_transmission_capacities — Function.

Plot transmission capacities.

source

EnergySystemModel.plot_storage — Function.

Plot storage.

source

EnergySystemModel.plot_storage_capacities — Function.

Plot storage capacities.

source

EnergySystemModel.plot_loss_of_load — Function.

Plot loss of load.

source

Utility

EnergySystemModel.equivalent_annual_cost — Function.

Equivalent annual cost (EAC)

Arguments

cost::Real: Net present value of a project.

n::Integer: Number of payments.

r::Real: Interest rate.

source